requirements.txt
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
Automat==20.2.0
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.1
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-crispy-forms==1.9.2
django-globals==0.3.2
django-js-asset==1.2.2
django-localflavor==3.0.1
django-utils-six==2.0
html5lib==1.1
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
mysqlclient==2.0.3
Pillow==8.1.0
pycparser==2.20
PyHamcrest==2.0.2
PyMySQL==1.0.2
Pyphen==0.10.0
python-dateutil==2.8.1
python-stdnum==1.15
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
tinycss2==1.1.0
Twisted @ file:///C:/paquetes/Twisted-20.3.0-cp39-cp39-win_amd64.whl
urllib3==1.26.2
validate-email==1.3
WeasyPrint==52.2
webencodings==0.5.1
whitenoise==5.2.0
zope.interface==5.2.0
